+++ projects/lite-form/src/public-api.ts
@@ -12,3 +12,4 @@ export * from './lib/lite-multi-select/lite-multi-select';
 export * from './lib/lite-radio/lite-radio';
 export * from './lib/lite-checkbox/lite-checkbox';
 export * from './lib/lite-date/lite-date';
+export * from './lib/lite-password/lite-password';